requirements.txt
langchain-core
langchain-openai
langchain-community
python-dotenv
pandas 
selenium
beautifulsoup4
openpyxl
openai